Changing theme of switches to iOS style
