+++ ControlCenterUi/src/Components/LightingDeviceSwitches.tsx
@@ -1,8 +1,7 @@
-import { Box, Grid } from "@mui/material";
-import Switch from "@mui/material/Switch";
+import { Box, Grid, styled } from "@mui/material";
+import Switch, { SwitchProps } from "@mui/material/Switch";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import React, { FC } from "react";
-
 interface ILightingDeviceSwitches {
   setBulbOneTarget(value: boolean): void;
   setBulbTwoTarget(value: boolean): void;
@@ -41,6 +40,57 @@ const formControlLabelStyle = {
   justifyContent: "center",
 };
 
+const IOSSwitch = styled((props: SwitchProps) => (
+  <Switch focusVisibleClassName=".Mui-focusVisible" disableRipple {...props} />
+))(({ theme }) => ({
+  width: 42,
+  height: 26,
+  padding: 0,
+  "& .MuiSwitch-switchBase": {
+    padding: 0,
+    margin: 2,
+    transitionDuration: "300ms",
+    "&.Mui-checked": {
+      transform: "translateX(16px)",
+      color: "#fff",
+      "& + .MuiSwitch-track": {
+        backgroundColor: theme.palette.mode === "dark" ? "#2ECA45" : "#65C466",
+        opacity: 1,
+        border: 0,
+      },
+      "&.Mui-disabled + .MuiSwitch-track": {
+        opacity: 0.5,
+      },
+    },
+    "&.Mui-focusVisible .MuiSwitch-thumb": {
+      color: "#33cf4d",
+      border: "6px solid #fff",
+    },
+    "&.Mui-disabled .MuiSwitch-thumb": {
+      color:
+        theme.palette.mode === "light"
+          ? theme.palette.grey[100]
+          : theme.palette.grey[600],
+    },
+    "&.Mui-disabled + .MuiSwitch-track": {
+      opacity: theme.palette.mode === "light" ? 0.7 : 0.3,
+    },
+  },
+  "& .MuiSwitch-thumb": {
+    boxSizing: "border-box",
+    width: 22,
+    height: 22,
+  },
+  "& .MuiSwitch-track": {
+    borderRadius: 26 / 2,
+    backgroundColor: theme.palette.mode === "light" ? "#E9E9EA" : "#39393D",
+    opacity: 1,
+    transition: theme.transitions.create(["background-color"], {
+      duration: 500,
+    }),
+  },
+}));
+
 const LightingDeviceSwitches: FC<ILightingDeviceSwitches> = (
   props
 ): JSX.Element => {
@@ -54,7 +104,7 @@ const LightingDeviceSwitches: FC<ILightingDeviceSwitches> = (
           <FormControlLabel
             style={formControlLabelStyle}
             control={
-              <Switch
+              <IOSSwitch
                 onChange={(event) =>
                   props.setBulbOneTarget(event.target.checked)
                 }
@@ -68,7 +118,7 @@ const LightingDeviceSwitches: FC<ILightingDeviceSwitches> = (
           <FormControlLabel
             style={formControlLabelStyle}
             control={
-              <Switch
+              <IOSSwitch
                 onChange={(event) =>
                   props.setBulbTwoTarget(event.target.checked)
                 }
@@ -82,7 +132,7 @@ const LightingDeviceSwitches: FC<ILightingDeviceSwitches> = (
           <FormControlLabel
             style={formControlLabelStyle}
             control={
-              <Switch
+              <IOSSwitch
                 onChange={(event) =>
                   props.setLedStripTarget(event.target.checked)
                 }
